requirements.txt
aenum==3.1.15
aiohappyeyeballs==2.4.4
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
beautifulsoup4==4.12.3
black==25.1.0
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cloudscraper==1.2.71
Deprecated==1.2.18
distro==1.9.0
et_xmlfile==2.0.0
Flask==3.1.0
frozenlist==1.5.0
future==1.0.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
line-bot-sdk==3.14.5
lxml==5.3.0
MarkupSafe==3.0.2
MouseInfo==0.1.3
multidict==6.1.0
mypy-extensions==1.0.0
numpy==2.2.1
openai==1.61.1
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.6
playwright==1.49.1
propcache==0.2.1
PyAutoGUI==0.9.54
pydantic==2.10.6
pydantic_core==2.27.2
pyee==12.0.0
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc-core==11.0
pyobjc-framework-Cocoa==11.0
pyobjc-framework-Quartz==11.0
pyparsing==3.2.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytweening==1.2.0
pytz==2024.2
requests==2.32.3
requests-toolbelt==1.0.0
rubicon-objc==0.5.0
selenium==4.27.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tqdm==4.67.1
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3